Update Pause / Game Over Menu + Home Menu
Now by pressing the escape button, you can pause the game and restart the level or go back to the Home Menu. There is also a new scene for the Home Menu, where you can choose to go to the Level Test or to quit the game. And when you get 0 hit points, the Game Over Menu appears and you can choose to restart the level or go back to the Home Menu. If you die or if you pause the game, the time is frozen and gets back to normal when you continue or restart the game.

PS : There is no game over if you fall.